v1.2.0 (ModMenu Support)

+++ src/main/java/com/notcharrow/instantfishing/config/InstantFishingConfig.java
@@ -6,5 +6,5 @@ public class InstantFishingConfig {
 	public boolean randomLootTable = false;
 	public boolean randomItem = false;
 	public boolean overrideTreasureChance = false;
-	public int treasureChance = 100;
+	public double treasureChance = 100.0;
 }
